C if,else,while

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int i =  0123456789;
+	long int i =  0123456789;
 	
 	putchar('i');
 	putchar('\n');
